Move around PC-related tests to add in PC cast
